refactor(input): move input classes to utilcalc.core.model.input package
