+++ core/src/main/java/utilcalc/core/model/input/HeatingFeeInputs.java
@@ -1,4 +1,4 @@
-package utilcalc.core.in;
+package utilcalc.core.model.input;
 
 import java.util.List;
