+++ core/src/main/java/utilcalc/core/model/input/SectionInput.java
@@ -1,4 +1,4 @@
-package utilcalc.core.in;
+package utilcalc.core.model.input;
 
 import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
