+++ core/src/main/java/utilcalc/core/model/input/SectionInputs.java
@@ -1,4 +1,4 @@
-package utilcalc.core.in;
+package utilcalc.core.model.input;
 
 public interface SectionInputs {
     String getName();
